Properly detect VPN on state
There was a problem where when the VPN was turned from 'off' to 'on' the
signal would sometimes not figure it out. I am not exactly sure how
these changes fixed it (it seemed to fix itself during development) but
I believe it had to do with the request being global to the module.

Defining the request on a per-call basis resolved the issue.