Always attach to below the network menu
Previously we were using a hardcoded value of '4' to appear below the
network menu. Sadly this means that number would be wrong if the user
had other extensions installed that also modified this menu.

We now lookup the position of the network menu and place ourselves below
it. Special thanks to @andyholmes on #extensions:gnome.org for helping
me work through this one.