Setting to disable indicator icon if disconnected
Adds a new setting to disable the indicator icon when not connected to
the VPN.

This makes it a bit easier for me to realize that the VPN is disabled,
instead of using a greyed out icon, perhaps a bit harder to distinguish
from the connected icon for the untrained eyes.

Doesn't change defaults, so the disconnected icon is still show by
default if the show-icon is enabled.